+++ l10n_br_account_product/models/__init__.py
@@ -3,6 +3,7 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from . import res_partner
+from . import account_fiscal_position_template
 from . import l10n_br_account_product
 from . import res_company
 from . import l10n_br_tax_definition_company_product
